new updates on clients and register

+++ src/main/java/com/drnavalhabarbershop/resourceserver/web/dto/BarberRequest.java
@@ -21,7 +21,6 @@ public class BarberRequest {
     private Integer userType;
     private String about;
     private Person person;
-    private Integer rating;
- //   private MultipartFile file;
+    private Double rating;
     private Binary profilePicture;
 }
